Hand and BestHand moved into model namespace
